Fix infinite loop issue on wrong input
When user input is incorrect (letter or number outside of allowed
range, etc.), program no longer loops forever. This is due to call to
`Skip_Line` which "consumes" wrong input so it does not pollute input
buffer and user can enter a new entry. It was not the case before the fix.

`Case` block was moved inside the `Main_Menu_Input` block, so previous
correct choice is no longer printed on wrong input as it was the case
before the fix.